Started offline refactor
In order to get things to work in both online and offline modes, some areas will have to be rebuilt to work primarily from a Cache state.

This first stage has the Server provide all assets that will need to be cached by the SW. Currently the login view displays in both online and offline.