fix: Notes don't load after login
If a User was already logged in and loaded a page, the notes would display. This was working because the loading of the notes data was in the `onMount` (so it'd only execute once).
Moved that logic to a function that's now called by `setUserInfo` after it's determined that User data exists.