fix: Can't delete Group after name changed
If a User edited a Group name and then wanted to delete that Group (without refreshing the page), an error would be thrown.
Turns out the initial computed props weren't being recalculated when the exported props were being changed (so the DOM was visibly updating, but not some of it's `data` attributes).
Moved the computed props to a `$:` context so props are recalculated.