fix: Odd lint warning for type definitions in svelte-portal
The 'Info' message seemed to be coming from the 'linter' Atom package, but I couldn't find any info about it anywhere, and I've been using that module in other repos without issue.

Just copied over the module for now, it was a fairly basic component so it shouldn't be an issue.