ops: Dev command not running in order
I previously switched to using the `build` command with `&`. Doing that started running prep-dist and started the Server at the same time, instead of what it used to do which was - run prep-dist, then start the WP watcher and start the Server. So I'll have to live with the duplication for now.