fix: Can't add Tag on Mobile
Best I can tell, the Tag input was triggering the default keyhint behavior for the input because it doesn't have a name (by design), so the Android Soft Keyboard only provides a the option to go to the Next input.
Luckily there's an API to inform Soft Keyboards how to behave, and I could tell it to present the Enter button instead.