fix: Line numbers in code blocks not lining up
I had a styling conflict with the Tag component. It had to have a global style for `.tag` since I'm forced to use a `Wrap` component to have a conditional anchor or div and Svelte doesn't include styles if it doesn't detect it in the file.
The conflict was with Prism which uses '.tag' for the styling of things like HTML tags.

Just changed the name of Tag to NoteTag. I wanted something more general purpose but what can yuh do.