fix: Auto-removal of blank list item not working on Android
Not sure what the cause is, but the code in `getSelectedLines` is detecting that there's no whitespace after a list marker, even though there is. My best guess is that the Browser or perhaps the Keyboard is trimming the extra whitespace trying to be helpful. Or there's a bug in the Browser?
Either way, I updated the list item detection to allow for no trailing whitespace (there had to be at least one whitespace for it to know it was a list item). I tried some different use cases and it didn't seem to break the functionality.