feat: Prevent the appearance of the caret jumping around when text inserted
In some cases when a User hits ENTER, the WYSIWYG will insert some text (like a list item). In that case, if it's inserting prefixed whitespace, you'd see the caret for a split second at the beginning of the line, then hop to the end of the inserted text. This change gets rid of that little blip.