feat: Don't display notes UI until everything loaded
While working on tests I realized there was a race condition where things weren't fully loaded but the Search and Notes buttons were enabled which would cause errors because they'd try to parse User data that wasn't there.
Now I'm waiting for data, Svelte to update the DOM (via `tick`), and the theme CSS to finish loading before displaying the Notes UI.
Now there's no flicker when a non-default theme loads in, and the tests aren't as brittle (depending on console logs was a band-aid).